+++ routes/users_route.js
@@ -5,11 +5,19 @@ const UsersDAO = require("../DAO/UsersDAO");
 const udao = new UsersDAO()
 
 router.get("/", async (req, res, next) => {
-    res.json(await udao.getAll())
+    if (await udao.checkToken(req)) {
+        res.send(await udao.getAll())
+    } else {
+        res.sendStatus(401)
+    }
 })
 
 router.get("/:id", async (req, res, next) => {
-    res.json(await udao.get(req.params.id))
+    if (await udao.checkToken(req)) {
+        res.send(await udao.get(req.params.id))
+    } else {
+        res.sendStatus(401)
+    }
 })
 
 router.post("/", async (req, res, next) => {
@@ -18,14 +26,23 @@ router.post("/", async (req, res, next) => {
 })
 
 router.put("/:id", async (req, res, next) => {
-    console.log(req.body, req.params.id)
-    res.json(await udao.updateUser(req.params.id, req.body.password))
+    const decoded = await udao.checkToken(req)
+
+    if (decoded.id == req.params.id) {
+        res.json(await udao.updateUser(req.params.id, req.body.password))
+    } else {
+        res.sendStatus(401)
+    }
 })
 
 router.delete("/:id", async (req, res, next) => {
-    console.log(req.body, req.params.id)
-    res.json(await udao.delete(req.params.id))
-})
+    const decoded = await udao.checkToken(req)
 
+    if (decoded.id == req.params.id) {
+        res.json(await udao.delete(req.params.id))
+    } else {
+        res.sendStatus(401)
+    }
+})
 
 module.exports = router;
